# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Exam Project",
    "description": "Documentation Exam Project"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/statuses/": {
      "get": {
        "tags": [
          "Statuses"
        ],
        "description": "Returns all statuses. Only admin can view",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/statuses/{id}": {
      "get": {
        "tags": [
          "Statuses"
        ],
        "description": "Returns one role which has an id equals to the provided id in the url path. Only admin can view",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/statuses/create": {
      "post": {
        "tags": [
          "Statuses"
        ],
        "description": "Create a new status. Only admin has access.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "An object which contains a statusName",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/statuses/update/{id}": {
      "put": {
        "tags": [
          "Statuses"
        ],
        "description": "Updates a status. id of the target status must be provided in the url path. Only admin has access",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "An object which contains a field newName",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/statuses/delete/{id}": {
      "delete": {
        "tags": [
          "Statuses"
        ],
        "description": "Deletes a specific status. status's id must be provided in the url path. Only admin!",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateStatus": {
      "type": "object",
      "properties": {
        "statusName": {
          "type": "string",
          "example": "In-progress"
        }
      },
      "required": [
        "statusName"
      ]
    },
    "UpdateStatus": {
      "type": "object",
      "properties": {
        "newName": {
          "type": "string",
          "example": "Completed"
        }
      },
      "required": [
        "newName"
      ]
    },
...
  }
}